[Sidebar] create dedicated Sidebar folder and refactor UserIcon
