+++ styles/components/Sidebar/userIcon.styles.tsx
@@ -1,0 +1,23 @@
+import styled from 'styled-components';
+
+/**
+ * Circular header image container
+ */
+export const ImageContainer = styled.div`
+  width: 100px;
+  height: 100px;
+  border-radius: 100%;
+  overflow: hidden;
+  background-color: white;
+  margin: auto;
+`;
+
+export const Name = styled.div`
+  margin: auto;
+  text-align: center;
+  color: white;
+  font-style: normal;
+  font-weight: 400;
+  font-size: 20px;
+  line-height: 30px;
+`;
